requirement.txt
requests
beautifulsoup4
gspread
oauth2client
python-dotenv